==> appaloosa/05_extract_EPIC_IDs_for_LC_search.ipynb <==
import pandas as pd
import numpy as np

# Pleiades

loc = 'stars_shortlist/Pleiades/results/run_06/catalog'
df1 = pd.read_csv('{}/Rebull_norotate.tsv'.format(loc), sep='\t',skiprows=55)
df1.head()

df2 = pd.read_csv('{}/Rebull_rotate.tsv'.format(loc), sep='\t',skiprows=75)
df2.head()

mergecols = list(set(df1.columns.values.tolist()) & set(df2.columns.values.tolist()))
df3 = df2.merge(df1, on=mergecols, how='outer')
df3.head()

df6 = pd.read_json('{}/panstarrs_search.json'.format(loc),encoding="utf8",lines=True,chunksize=1000)
for line in df6

df3 = df3.drop(['Fig','_RAJ2000','_DEJ2000'], axis=1)
df3.to_csv('{}/Pleiades_init.csv'.format(loc))

#  M44

loc = 'stars_shortlist/M44/results/run_06/catalog'
df4 = pd.read_csv('{}/Douglas.tsv'.format(loc), sep='\t',skiprows=103)
df4 = df4.rename(index=str,columns={'_RA':'RAJ2000','_DE':'DEJ2000'})
df4 = df4.drop(['_RAJ2000','_DEJ2000'],axis=1)
df4.head()

df4.to_csv('{}/M44_init.csv'.format(loc))

df4.shape

# M67

loc = 'stars_shortlist/M67/results/run_06/catalog'
df5 = pd.read_csv('{}/Gonzalez.tsv'.format(loc), sep='\t',skiprows=59)
df5 = df5.rename(index=str,columns={'_RA':'RAJ2000','_DE':'DEJ2000'})
df5 = df5.drop(['_RAJ2000','_DEJ2000','SimbadName'],axis=1)
df5.head()

df5.Class = df5.Class.str.strip()
df5 = df5[df5.Class.isin(['M','SM','BM']) ]
df5.to_csv('{}/M67_init.csv'.format(loc))

# Any cluster: 
## Merge Pan-STARRS with K2 and initial sample from Rebull; Douglas; Gonzalez

cluster = 'Pleiades'
test =  'run_06'
campaign = 4
loc = 'stars_shortlist/{}/results/{}/catalog'.format(cluster,test)
df5 = pd.read_csv('{}/{}_init_PS1.csv'.format(loc,cluster))
#df5 = df5.rename(index=str, columns={})
df6 = pd.read_csv('{}/{}_init_K2search.csv'.format(loc,cluster),skiprows=[1])
df6 = df6.rename(index=str, columns={'K2 ID':'EPIC', 'J Mag':'J', 'H Mag': 'H',  'K Mag': 'K'})
df6 = df6[df6['Campaign'] == campaign]
grouped = df5.groupby('EPIC')
df7 = grouped.apply(lambda x: x[x.angDist == x.angDist.min()])
print(df7.shape)
df8 = df7.merge(df6, how='right', on='EPIC')
print(df8.shape)

df8.columns.values

from PS_to_SDSS import wrap
df8.to_csv('{}/{}_parameter.csv'.format(loc,cluster))
wrap(cluster, test)

df8 = pd.read_csv('{}/{}_parameter.csv'.format(loc,cluster))
df8.columns.values
