==> appaloosa/11B_systematics_stats.ipynb <==
%matplotlib inline
import warnings
warnings.filterwarnings("ignore")
import time
import os
import math
import numpy as np
import pyke as pk
import pandas as pd
import k2flix
import matplotlib.pyplot as plt

def collect(oid_list,test,cluster):
    tpfs, flux, flares = dict(), dict(), dict()
    for oid in oid_list:
        print(oid)
        flares[oid] = pd.read_csv('stars_shortlist/{}/results/{}/{}_flares.csv'.format(cluster,test,oid),)
                              #   usecols=['istart','istop','istart_no_sys_no_thruster','istop_no_sys_no_thruster'])
        flux[oid] = pd.read_csv('stars_shortlist/{}/results/{}/{}_flux.csv'.format(cluster,test,oid))#,
                                #usecols=['flux_gap','time','flux_model'])
    return flux, flares


#---------------------------------------------------------------------------
# set parameters here
cluster = 'M67'

oid_list = pd.read_csv('stars_shortlist/{0}/{0}_parameter.csv'.format(cluster))#,names=['EPIC'])
oid_list = oid_list.EPIC.tolist()
test = 'run_05'
#----------------------------------------------------------------------------
print(os.getcwd())
print(len(oid_list))
flux, flares = collect(oid_list, test,cluster)


l1, l2 = [], []
for oid in oid_list:
    try:
        l1.append(len(flares[oid].istart.dropna()))
        l2.append(len(flares[oid].istart_no_sys_no_thruster.dropna()))
    except AttributeError:
        continue
c_flares = pd.DataFrame({'before':l1,'after':l2},index=oid_list)
c_flares['difference'] = (c_flares.after - c_flares.before)/c_flares.before
c_flares.describe()

flares[oid_list[2]].head()

flux[oid_list[0]].head()

def flag_df(flux):
    
    '''
    
    Plug in here any measure of the results that contains information for your interpretation.
    
    Input:
    
    flux - DataFrame that contains all previous processing.
    
    Return:
    
    flags - Table with customized inputs
    
    '''
    index = [1,8192,524288]
    columns = ['>',
               '>_after_thruster_removal',
               '>_after_systematics_removal',
               '>_left_overall',
               '>_coinciding_with_flare_candidates',
               '>_coinciding_with_clean_flare_candidates',
               '>_relative_coinciding_with_flare_candidates',
               '>_relative_coinciding_with_clean_flare_candidates',
              ]
    flags = pd.DataFrame(index=index, columns=columns)
    for flag in index:
        flags[columns[0]].loc[flag] = flux[flux.tpf_flags >= flag].shape[0]
        flags[columns[1]].loc[flag] = flux[flux.new_tpf_flags >= flag].shape[0]
        flags[columns[2]].loc[flag] = flux[flux.tpf_flags_wo_systematics >= flag].shape[0]
        flags[columns[3]].loc[flag] = flux[(flux.tpf_flags_wo_systematics >= flag) & (flux.new_tpf_flags >= flag)].shape[0]
        flags[columns[4]].loc[flag] = flux[(flux.isflare == True) & (flux.tpf_flags >= flag)].shape[0]
        flags[columns[5]].loc[flag] = flux[(flux.isflare_no_sys_no_thruster == True) & (flux.tpf_flags >= flag)].shape[0]
        flags[columns[6]].loc[flag] = np.around(flux[(flux.isflare == True) & (flux.tpf_flags >= flag)].shape[0]/flux.shape[0],3)
        flags[columns[7]].loc[flag] = np.around(flux[(flux.isflare_no_sys_no_thruster == True) & (flux.tpf_flags >= flag)].shape[0]/flux.shape[0],3)

    return flags.T

for o in oid_list[:2]:
    print('\n',o)
    print(flag_df(flux[o]))

dfi = flag_df(flux[oid_list[0]])
for o in oid_list[1:]:
    dfi = dfi.add(flag_df(flux[o]))
df_av = dfi/len(oid_list)
#df_av.to_csv('results/{0}/plots/{1}_{0}_average_flare_stats.csv'.format(test,cluster))

df = flag_df(flux[oid_list[0]]).stack()
df2 = flag_df(flux[oid_list[1]]).stack()
summary = pd.DataFrame({oid_list[0]: df})
for o in oid_list[1:]:
    ser = flag_df(flux[o]).stack()
    ser.name = o
    summary = summary.join(ser)
#summary.to_csv('results/{0}/plots/{1}_{0}_all_flare_detail.csv'.format(test,cluster))

std = pd.DataFrame({'maximum':summary.max(axis=1),
                    'minimum': summary.min(axis=1),
                    'average':summary.mean(axis=1),
                    'std dev':summary.std(axis=1)})
#std.to_csv('results/{0}/plots/{1}_{0}_all_flags_stats.csv'.format(test,cluster))

df_av


std
