==> appaloosa/14_wrapup.ipynb <==
import pandas as pd
%matplotlib inline
cl = ['Pleiades','M44']
df = pd.DataFrame()
for c in cl:
    d = pd.read_csv('stars_shortlist/{0}/results/run_06/catalog/{0}_parameter.csv'.format(c),
                    usecols=['EPIC','todrop','whydrop'])
    d['cluster'] = c
    df = df.append(d, ignore_index=True)
    
r = pd.DataFrame(columns=['note','cluster','targets'])  
for name, d in df.groupby('cluster'):
    c = d.whydrop.value_counts().reset_index().rename(index=str, columns={'whydrop':'targets',
                                                                          'index':'note'})
    c['cluster'] = name
    c.note[c.note=='    '] = 'good'
    c.note[c.note=='    noLum'] = 'missing photometry'
    r = r.merge(c, on=['cluster','note','targets'],how='outer')

r.set_index('cluster')
